Add a PYTHONREVERSEDICTKEYORDER environment variable.
As part of the reproducible builds [0] initiative we would like to find
any code that is relying on non-determinstic or undefined behaviour. In
Python 2.x, this includes dict ordering.

It would therefore be convenient to be able to specify an environment
variable that would reverse the usual ordering; we would then run a build
of an arbitrary package both with and without this flag and compare the
resulting binaries. We already run such a comparison framework, so
specifying this variable would be trivial for us.

Note that we are only reversing the order when this variable is set; we
do not need to sort first as we are only concerned with flushing out
differences. This patch is controversial for Python >= 3.6 as the dict
ordering is defined to be consistent there in some situations.

This patch changes the behaviour of:

  * for x in d:
  * d.popitem()
  * d.items() & d.iteritems()
  * d.values() & d.itervalues()
  * print(d)
  * repr(dx)
  * PyDict_Next & _PyDict_Next

It does not change the behaviour of collections.OrderedDict.

 [0] https://reproducible-builds.org/

Signed-off-by: Chris Lamb <chris@chris-lamb.co.uk>